refactor: share THEME_COLOR_KEYS via a common/ dir across the build boundary
THEME_COLOR_KEYS (the 24 xterm ITheme keys a `.mulmoterminal.json` colors block may
override) was duplicated verbatim between the server config schema and the client
dir-config parser — the largest clone in the repo (25 lines). The two builds have no
shared directory (tsconfig.server = server/**, tsconfig.app = src/**), which is why it
was copied; a key added to one list but not the other would be silently accepted on one
side and dropped on the other.

Add a common/ dir for pure code shared across the boundary and move the constant there.

- common/themeColors.ts: the single source of truth.
- Both tsconfigs include common/**; package.json `files` ships common/ (the server runs
  from source via tsx, so the published package needs it, not just the bundled client).
- knip.json project covers common/ so it's watched for dead exports too.

Verified the boundary end-to-end: server typecheck + client vue-tsc + vite build all
resolve it, tsx resolves it at runtime (dirColorsField still strips unknown keys), and
`npm pack` includes common/themeColors.ts.

Deliberately NOT unified: the server and client DirConfig types (jscpd flags their first
~13 fields as a clone) are intentionally different — the server keeps the raw `sound`
path and buttons/chips/skills, the client gets only `hasSound` (the path stays
server-side by design). Merging them would break that boundary.
