+++ package.json
@@ -22,6 +22,7 @@
   },
   "files": [
     "bin/",
+    "common/",
     "dist/",
     "server/",
     "plugins/",
