+++ tsconfig.app.json
@@ -11,5 +11,5 @@
     "erasableSyntaxOnly": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.vue", "plugins/**/*.ts", "plugins/**/*.vue"]
+  "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.vue", "plugins/**/*.ts", "plugins/**/*.vue", "common/**/*.ts"]
 }
